split data service file into more files
